# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Servanda Office API
  version: 1.0.0
  description: |
    Multi-Tenant LegalTech Platform API for contract creation and management.

    **Core Flow:** Template selection -> Guided Q&A Interview -> Conflict-free Contract -> DOCX/ODT Export.

    **Architecture Decisions:**
    - ADR-001: Shared DB with PostgreSQL Row-Level Security for tenant isolation
    - ADR-002: Immutable versioning for Clauses/Templates; ContractInstance pins versions
    - ADR-003: Async export via pgboss job queue with S3 storage
    - ADR-004: Server-side ODT conversion via LibreOffice headless

    **Authentication:**
    - Production: Bearer JWT (OIDC/SAML)
    - Development: Header-based tenant context (x-tenant-id, x-user-id, x-user-role)

    **Pagination:** All list endpoints support `page` and `pageSize` query parameters.
    Default page size: 20, maximum: 100.

    **Status Workflow (Clause/Template Versions):**
    `draft` -> `review` -> `approved` -> `published` -> `deprecated`
    (review -> draft and approved -> draft are also valid transitions)
  contact:
    name: Team 01 - Product Architecture
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server (v1 canonical prefix)
  - url: http://localhost:3000/api
    description: Legacy route prefix (backward-compatible)
security:
  - BearerAuth: []
  - DevHeaders: []
...
paths:
...
  /content/clauses:
    post:
      tags: [Content]
      summary: Create a clause
      description: |
        Creates a new clause entity. Requires `admin` or `editor` role.
        The clause is created without any versions; add versions separately.
      operationId: createClause
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClauseRequest'
      responses:
        '201':
          description: Clause created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clause'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Content]
      summary: List clauses
      description: Returns a paginated list of all clauses in the current tenant.
      operationId: listClauses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of clauses
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Clause'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /content/clauses/{id}:
    get:
      tags: [Content]
      summary: Get clause with versions
      description: Returns a single clause by ID including all its versions ordered by version number descending.
      operationId: getClause
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Clause with versions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Clause'
                  - type: object
                    properties:
                      versions:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClauseVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/clauses/{id}/versions:
    post:
      tags: [Content]
      summary: Create a new clause version
      description: |
        Creates a new immutable version of a clause (ADR-002).
        Version number is auto-incremented. Initial status is `draft`.
        Optionally includes conflict/dependency rules for clause consistency validation.
      operationId: createClauseVersion
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClauseVersionRequest'
      responses:
        '201':
          description: Clause version created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClauseVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/clauses/{id}/versions/{vid}/status:
    patch:
      tags: [Content]
      summary: Transition clause version status
      description: |
        Transitions the status of a clause version following the workflow:
        `draft` -> `review` -> `approved` -> `published` -> `deprecated`.
        Also allows `review` -> `draft` and `approved` -> `draft`.

        When transitioning to `published`, publishing gates are validated automatically.
        If gates fail, the transition is rejected with a 409 Conflict.
      operationId: transitionClauseVersionStatus
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusTransitionRequest'
      responses:
        '200':
          description: Status transitioned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClauseVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid status transition or publishing gates failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /content/clauses/{id}/versions/{vid}/publishing-gates:
    get:
      tags: [Content]
      summary: Check clause version publishing gates
      description: |
        Pre-flight check that validates all publishing gates for a clause version
        without actually transitioning the status. Returns gate results with
        pass/fail status and severity.
      operationId: checkClausePublishingGates
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/VersionIdParam'
      responses:
        '200':
          description: Publishing gate validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishingGateResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/clauses/batch-content:
    post:
      tags: [Content]
      summary: Batch-fetch clause version content
      description: |
        Fetches content for multiple clause versions in a single request.
        Used by the Live-Preview-Panel in the Contract Builder UI.
        Minimum 1, maximum 50 clause version IDs per request.
      operationId: batchClauseContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchClauseContentRequest'
      responses:
        '200':
          description: Batch clause version content
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClauseVersionContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =========================================================================
  # CONTENT - Templates
  # =========================================================================
  /content/templates:
    post:
      tags: [Content]
      summary: Create a template
      description: |
        Creates a new template entity. Requires `admin` or `editor` role.
        The template is created without versions; add versions separately.
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Content]
      summary: List templates
      description: Returns a paginated list of all templates in the current tenant.
      operationId: listTemplates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of templates
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Template'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /content/templates/{id}:
    get:
      tags: [Content]
      summary: Get template with versions
      description: Returns a single template by ID including all its versions ordered by version number descending.
      operationId: getTemplate
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Template with versions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Template'
                  - type: object
                    properties:
                      versions:
                        type: array
                        items:
                          $ref: '#/components/schemas/TemplateVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production authentication via OIDC/SAML JWT token.
        Include tenant claim in the JWT payload.
    DevHeaders:
      type: apiKey
      in: header
      name: x-tenant-id
      description: |
        Development-mode authentication via headers.
        Requires all three headers:
        - `x-tenant-id` (UUID) - Tenant identifier
        - `x-user-id` (UUID) - User identifier
        - `x-user-role` (admin|editor|user) - User role

  # =========================================================================
  # Parameters
  # =========================================================================
...